Revert reduce-run-commands
Revert change to reduce-run-commands to use a local batch file to run
PSL REDUCE, which is the only way I can get it to be able to load
packages.  Revert use of cmd.exe instead of cmdproxy.exe.  Introduce
option reduce-run-MSWin-drives and find all local drives on MS Windows
automatically.  Rename reduce-installation-directory to
reduce-run-installation-directory, and use reduce-run-MSWin-drives to
assign a more reliable default value to
reduce-run-installation-directory.