reduce-run.el
run-reduce: command name defaults to the last value used.
switch-to-reduce: create a new window on the right.
reduce--wait-for-prompt: revert to safer implementation.

But switch-to-reduce still hangs when starting a new REDUCE process.
Emacs recovers after ^G.  The problem seems to be in
reduce--wait-for-prompt, but I can't see what it is!