Version 1.8alpha: revise sexp commands
Update version to 1.8alpha.  The commands reduce-forward-sexp and
reduce-backward-sexp no longer accept an argument and now skip any
comments or white space, but not terminators, between point and the
"balanced expression".  Add the command reduce-kill-sexp to kill one
"balanced expression" either forwards or, with any argument,
backwards, bound to `C-M-k`.