Parser -- use vector
Make reduce--comment-seq a vector or nil if there are no comment
statements, which is an important special case.  Use mapc instead of
seq-do.  Revise the bisection algorithm, which now seems to work and
examines uniformly around log2 n of the n comment statement intervals,
rather than the average n/2 for a linear search.