Make thing-at-point fix more robust
Introduce function reduce--update-bounds-of-thing-at-point run by the
hook buffer-list-update-hook to control the bounds-of-thing-at-point
property of symbol.  This should mean that
reduce--bounds-of-symbol-at-point is used only in REDUCE mode.  But
there could be a clash with other modes, in which case I would need to
handle this in a more sophisticated way by saving and restoring this
property (which is nil by default).  I will hope that any other mode
that needs to set this property will do so in the way that I do, which
should be reliable.