Version 1.11alpha: Further improve Imenu support
Revert to default name of “Index” for Imenu menu on menubar.  Rename
reduce-imenu to reduce-imenu-add and change its default value to t
(since modern computers seem to be fast enough to build an imenu).
Rename reduce--imenu-done to reduce--imenu-added.  Rename
reduce--imenu-add-to-menubar to reduce--imenu-add-menubar-index and
call imenu-add-menubar-index instead of imenu-add-to-menubar.  Update
reduce--imenu-create-index to include variables and exclude empty
sub-menus.  Update reduce-show-proc-regexp not to use
reduce-imenu-generic-expression, which no longer exists, but this
needs further work.  Update documentation.