Fix CSL REDUCE on Linux
Specify a suitable terminal type so that CSL REDUCE responds
appropriately to interrupts, which with the default dumb terminal it
does not.  Can perhaps do this more elegantly in the code that starts
the process at a later date!  Only remove ansi-color-process-output on
MS Windows; it is necessary for CSL REDUCE on Linux and probably
redundant on MS Windows.