Rename reduce-run-installation-directory
Rename `reduce-run-installation-directory` to
`reduce-root-dir-file-name` and make it a directory file name rather
than a directory name, i.e. remove the final directory separator, so
that it can be used as the value of $reduce (which must be lower
case).  Move the definition of `reduce-packages-directory` up to join
the other options and move the hook definitions to the end of this
section.