Better support for multiple REDUCE installations
Allow $reduce to be used separately for each command so that
configuring multiple REDUCE installations is easier and more
consistent.  Always use $reduce, if set, to set the "reduce"
environment variable.  This requires small changes to the details of
reduce-root-dir-file-name and reduce-run-commands.  I would really
like to force the value of reduce-root-dir-file-name, etc., to be an
absolute directory file name but I can't get this to work!

This version is transitional and not fully working.  I still need to
make package loading work separately for each command.